requirements/base.txt
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.12.2